Fix: restore tracking of send modals (#2185)
